Minor changes for aquire_scihub check_scihub functions and scripts files

+++ inst/scripts/BSCI.js
@@ -7,7 +7,7 @@ function setup() {
             }],
         output: { // this defines the output image type
             bands: 1,
-            sampleType: "UINT8"
+            sampleType: "FLOAT32"
         }
     };
 }
