+++ inst/scripts/STR.js
@@ -8,7 +8,7 @@ function setup() {
             }],
         output: { // this defines the output image type
             bands: 1,
-            sampleType: "UINT8"
+            sampleType: "FLOAT32"
         }
     };
 }
